requirements.txt
Flask==2.2.5
tensorflow==2.18.0
tensorflow-hub==0.16.0
Pillow==9.4.0
numpy>=1.24.3
